Moved the code responsible for font size from the application class to FontSizeProvider
