+++ collect_app/src/main/java/org/odk/collect/android/activities/FormEntryActivity.java
@@ -141,6 +141,7 @@ import org.odk.collect.android.utilities.ApplicationConstants;
 import org.odk.collect.android.utilities.DestroyableLifecyleOwner;
 import org.odk.collect.android.utilities.DialogUtils;
 import org.odk.collect.android.utilities.FileUtils;
+import org.odk.collect.android.utilities.FontSizeProvider;
 import org.odk.collect.android.utilities.FormNameUtils;
 import org.odk.collect.android.utilities.ImageConverter;
 import org.odk.collect.android.utilities.MediaManager;
@@ -367,7 +368,7 @@ public class FormEntryActivity extends CollectAbstractActivity implements Animat
             showPreviousView();
         });
 
-        questionFontSize = Collect.getQuestionFontsize();
+        questionFontSize = FontSizeProvider.getQuestionFontsize();
 
         if (savedInstanceState == null) {
             mediaLoadingFragment = new MediaLoadingFragment();
@@ -2087,8 +2088,8 @@ public class FormEntryActivity extends CollectAbstractActivity implements Animat
     }
 
     private void adjustFontSize() {
-        if (questionFontSize != Collect.getQuestionFontsize()) {
-            questionFontSize = Collect.getQuestionFontsize();
+        if (questionFontSize != FontSizeProvider.getQuestionFontsize()) {
+            questionFontSize = FontSizeProvider.getQuestionFontsize();
             refreshCurrentView();
         }
     }
